[♻️ Refactor] 컴포넌트 분리를 위한 폴더 구조 변경 (#none)
