+++ app/(tabs)/running/tracking.tsx
@@ -1,6 +1,6 @@
-import CountDown from "@/components/board/CountDown";
-import RunDataBoard from "@/components/board/RunDataBoard";
-import TipBoard from "@/components/board/TipBoard";
+import CountDown from "@/components/board/RunBoard/CountDown";
+import RunDataBoard from "@/components/board/RunBoard/RunDataBoard";
+import TipBoard from "@/components/board/RunBoard/TipBoard";
 import RunControlButton from "@/components/button/RunControlButton";
 import GoogleMap from "@/components/map/GoogleMap";
 import useNonNavbar from "@/hooks/use-non-navbar";
