+++ components/board/HomeDashBoard/WalkScoreBoard.tsx
