+++ components/board/RunBoard/RunResultBoard.tsx
@@ -2,7 +2,7 @@ import { useRunStore } from "@/store/useRunStore";
 import { formatTime } from "@/util/formatTime";
 import React from "react";
 import { Text, View } from "react-native";
-import { HStack } from "../ui/hstack";
+import { HStack } from "../../ui/hstack";
 
 const RunResultBoard = () => {
   const { runData } = useRunStore();
